backend/creative-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.53.0
anthropic==0.39.0
httpx==0.27.2
replicate>=1.0.0
graphviz==0.20.3
Pillow==11.0.0
structlog==24.1.0
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.1.0

# OpenTelemetry tracing
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

